image: Fix nginx TLS config and gencerts script
This mostly fixes the bug described in commit 3444d0e. The nginx config didn't
work on a fresh instance; the cert and key files didn't exist. This splits the
nginx TLS and non-TLS configs into separate files and has gencerts link the
TLS config when it first runs successfully. Now gencerts script relies on a
magic comment in each nginx config file to tell it what domain to renew.